Update LogoBranding component
