+++ src/components/LogoBranding/index.js
@@ -1,0 +1,3 @@
+import LogoBranding from './LogoBranding';
+
+export default LogoBranding;
